Directly wrapping LLVM binary and memory instr(s)
This way an explicit decision about the instruction being named or
unnamed has to be made when adding that instruction to the basic block's
stack. The wrapping of instructions is simple, with binary instructions
only requiring two operands (with default options specified
implicitely).

The decision of naming an instr is left to the `Blocks` module which
will usually try to use the LPVM's given name, since it is already in a
SSA form. This makes the naming scheme more directly related to LPVM
instead of using unnamed incremental suffixes.