Added collection of prototypes from imports
This is done before translation of the ProcDefs so that a PrimCall
argument can be matched with its counterpart in the
prototype (PrimParam). The PrimParam has information which determines if
it is needed or not. Using this label, a matching PrimArg can be
filtered out of the call and extern declaration for that Procedure.