Fixed: Searching for types + TypeRepr lookup
When searching for a type in a module, the parent mods are also searched
if the search returns an empty set. This makes types recognizable in the
submod the type was declared in.

When doing a lookup for the TypeRepresentation, the searched modules are
not assumed to have been already loaded (could be under compilation
instead too). The search precendence followed now:
Interface --> Implementation --> Parent mod...