Fixed LLVM not finding value in slotcalculator
When a call to a void returning function is made, there is no actual
return operand or LHS value for the call. The previous erroneous
behaviour tried to return this value at the end of the branched block.

Now, a void operand is filtered to just be Nothing, so instead of
returning a void op, only a void value will be returned (not tied to an
operand which does not exist).