Fixed temp file IO clash error
Using the same file in the temp directory to store the bytestring for every
module which is fed to ld's `-sectcreate' option was causing some object
files to be created with an empty bytestring.

Using a different temp file for different modules (based on the module
name) fixes this synchronisation issue.