Implement last call modulo construction optimization (#279)
* initial exploration - identify last call and prims before/after

* handle PrimHigher variant

* move prims which are trivially moveable before last call, decide when optimization is safe to apply

also added a few bits of sample code I've been using to test what I've written. I'm sure I'll get rid of it eventually but am putting it in the compiler repo for now

* added FlowOutByReference, started modifying proc to use FlowOutByReference

* working `append` transformation, but test-cases in a bad shape

* add alloca lpvm instruction

* remove examples

* fix build

* fix compile, add FlowTakeReference

* LastCallAnalysis only reorders, annotates procs

Now I need to update codegen to work with this new LPVM post-analysis pass

* get rid of Pointer, PointerRep, use PrimFlow more in codegen

* compiling append example with new codegen, added print_list function

* only emit musttail on directly recursive functions calls

there is still a bug: we use `musttail` on functions which return tuples, which get
unpacked/repacked by our codegen - thus LLVM complains: musttail call must
precede a ret with an optional bitcast

* make BodyBuilder / Transform work with new flow types

BodyBuilder / Transform was the culprit that made the multi-specz function bodies empty. I've fixed that so the codegen works nicely, but it involved reverting the ordering of call/append back to normal ordering...

* add pass which coerces FlowOut into FlowOutByReference where needed

* remove "FlowOutByReferenceShim", replace with upgrade

* cleanup diff ready for PR

* furhter cleanup, add comments, fix test cases

* add preassignOutputs back again, more cleanup

* revert change to emit - put in separate PR

* get rid of `tail` LPVM attribute, instead infer from context, more codegen cleanup

* more bugfixes to codegen

* update test cases - need to look over them again thoroughly for regressions

* explain why tail call kind was chosen

* add example where higher order function calls last-call optimised append

also abort last call analysis if the proc we are analysis is a ClosureProc variant

* remove unused import

* fix test cases, make list.print take a "printer" argument

* more test-cases changed