write outByRef outputs directly into structure
this is a further optimization to existing TCMC optimisation, which tries to optimize the caller of these tail-call-optimized procs. We identify the pattern of `alloca` -> call -> `load` -> mutate() which occurs quite often. Instead we calculate the address and then call passing in the final address to write to (i.e.: writing *directly into the structure*)