Implement basic string string interpolation
Allow $varname in string literals, scanning them as the string before
the $varname, a string concatenation (,,), a call to `fmt` with
varname as argument, another string concatenation, and finally the
string after the $varname.  This is all handled in the scanner.

Also wrote a `fmt` function for each of the atomic primitive types.
We can't really have a compatible `fmt` function for parametric types,
because it would need to take the `fmt` function for each type
parameter as another input, so the interface would be wrong.  We may
need type classes to solve this.