Revert on failure (#419)
* Fix #252: forbid reification of tests with outputs

* Move determinism into Unbrancher monad

* Code doc improvements; delete reduntant calls

* Move alternative (else) code into Unbrancher monad

* Move test sense into Unbrancher monad

* Track list of variables to backtrack on failure
Also doc and logging improvement.
Discovered that the brVars list of defined variables is not
always correct as bodies are unbranched, because the branches
of a conditional are unbranched before the condition.  This
defeats this strategy for tracking the variables to save and
restore.

* Ignore LOG files stored here

* Back out changes to Unbranch, as that won't work.

* Move detertminism into Typed monad

* Refactor:  move proc name, module, and src pos into Typed monad

* Fix #412
Foreign llvm {test} instructions are given one extra Boolean output
argument, which determines the success of the test.  Foreign other
language {test} instructions are given an extra int output argument,
which is interpreted as success iff it is different from 0 (following
C conventions).  To implement this, TestBool instructions now allow
the argument to be an int, in which case it is compared to 0 to
determine the test outcome.

* Refactor mode checking to use Moded monad

* Roll back assignments in failing conditional test

* Backtrack bindings in failing disjuncts; add test

* Remove accidentally added files