Handle conversions better Just use bitcast for automatic integer <-> float converstions, to preserve all information.
Track assigned LLVM variable types, so we can
convert when the value is subsequently used as a
different type.  This happens for generics, when a
typed value is supplied for a generic value.  This
really isn't ideal, but the type checker can give
the same variable different types in this case.